Add boiler-plate code for app containing login flow with logged-in pages set up
Add the code thus far. All the app can do right now is just contact
the schedulearn api and login. After someone logs in, their token
is stored as a SharedPreference, as well as a boolean value that
indicates the user is logged in. This boolean is checked on the
homepage and if it's true, the app redirects you to the main
activities. The main activities contain a bottom navigation bar and
a fragment so you can switch fragments when you go to the different
navs